// src/screens/checkout/checkoutAddress.selectors.types.ts
export type CheckoutAddressSelectors = {
  readonly screen: string;
  readonly fullNameField: string;
  readonly address1Field: string;
  readonly address2Field: string;
  readonly cityField: string;
  readonly stateRegionField: string;
  readonly zipField: string;
  readonly countryField: string;
  readonly toPaymentButton: string;
};